Widen channel color to BIGINT and fix legacy migration stamping.
Lets Flutter ARGB values fit in Postgres and still apply later Alembic revisions.

+++ cloud/app/models.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import datetime, timezone
 from typing import List, Optional
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
+from sqlalchemy import Boolean, BigInteger, DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .db import Base
@@ -83,8 +83,8 @@ class ChannelRow(Base):
     default_allow_seek: Mapped[bool] = mapped_column(Boolean, default=True)
     sort_order: Mapped[int] = mapped_column(Integer, default=0)
     source_type: Mapped[str] = mapped_column(String(40), default="youtubePlaylist")
-    # Signed int32 form of Flutter ARGB (0xFF42A5F5).
-    color: Mapped[int] = mapped_column(Integer, default=-12409355)
+    # Flutter ARGB fits in BIGINT; import also clamps to signed int32 for safety.
+    color: Mapped[int] = mapped_column(BigInteger, default=-12409355)
     videos: Mapped[List["VideoRow"]] = relationship(
         back_populates="channel",
         cascade="all, delete-orphan",
